Rename dockingTask to docking
