+++ frontend/client/visualize/docking/main.tsx
