+++ frontend/client/visualize/docking/right-panel.css
