+++ frontend/client/visualize/docking/right-panel.tsx
