+++ frontend/client/visualize/docking/types.ts
