+++ frontend/client/visualize/docking/visualization-box.tsx
